# Conflicts:
#	apps/api/plane/authentication/provider/credentials/email.py

+++ apps/web/core/components/estimates/inputs/root.tsx
@@ -1,4 +1,9 @@
-import type { FC } from "react";
+/**
+ * Copyright (c) 2023-present Plane Software, Inc. and contributors
+ * SPDX-License-Identifier: AGPL-3.0-only
+ * See the LICENSE file for details.
+ */
+
 // plane imports
 import type { TEstimateSystemKeys } from "@plane/types";
 import { EEstimateSystem } from "@plane/types";
@@ -21,7 +26,7 @@ export function EstimateInputRoot(props: TEstimateInputRootProps) {
     case EEstimateSystem.POINTS:
       return (
         <EstimateNumberInput
-          value={value ? parseInt(value) : undefined}
+          value={value ? parseFloat(value) : undefined}
           handleEstimateInputValue={handleEstimateInputValue}
         />
       );
